refactor(ms2/api-images): extracted logic to helpers
