+++ src/management-system-v2/app/api/private/[environmentId]/processes/[processId]/images/[imageFileName]/route.ts
@@ -8,11 +8,10 @@ import {
   saveProcessImage,
 } from '@/lib/data/legacy/_process';
 import { NextRequest, NextResponse } from 'next/server';
-import stream from 'node:stream';
-import type { ReadableStream } from 'node:stream/web';
 import jwt from 'jsonwebtoken';
 
 import { TokenPayload } from '@/lib/sharing/process-sharing';
+import { invalidRequest, readImage } from '../../../../image-helpers';
 
 export async function GET(
   request: NextRequest,
@@ -102,57 +101,13 @@ export async function PUT(
     });
   }
 
-  const allowedContentTypes = ['image/jpeg', 'image/svg+xml', 'image/png'];
+  const isInvalidRequest = invalidRequest(request);
+  if (isInvalidRequest) return invalidRequest;
 
-  const contentType = request.headers.get('content-Type');
+  const readImageResult = await readImage(request);
+  if (readImageResult.error) return readImageResult.error;
 
-  if (!contentType || !allowedContentTypes.includes(contentType)) {
-    return new NextResponse(null, {
-      status: 400,
-      statusText: 'Wrong content type. Image must be of type JPEG, PNG or SVG.',
-    });
-  }
-
-  if (!request.body) {
-    return new NextResponse(null, {
-      status: 400,
-      statusText: 'No image was given in request',
-    });
-  }
-
-  const reader = stream.Readable.fromWeb(request.body as ReadableStream<Uint8Array>);
-  const chunks: Uint8Array[] = [];
-  let totalLength = 0;
-  for await (const chunk of reader) {
-    if (chunk) {
-      chunks.push(chunk);
-      totalLength += chunk.length;
-      if (totalLength > 2000000) {
-        // 2MB limit
-        reader.destroy(new Error('Allowed image size of 2MB exceeded'));
-        return new NextResponse(null, {
-          status: 413,
-          statusText: 'Allowed image size of 2MB exceeded',
-        });
-      }
-    }
-  }
-  // Proceed with processing if the size limit is not exceeded
-  const imageBuffer = Buffer.concat(
-    chunks.map((chunk) => Buffer.from(chunk)),
-    totalLength,
-  );
-
-  const fileType = await fileTypeFromBuffer(imageBuffer);
-
-  if (!fileType) {
-    return new NextResponse(null, {
-      status: 415,
-      statusText: 'Can not store image with unknown file type',
-    });
-  }
-
-  await saveProcessImage(processId, imageFileName, imageBuffer);
+  await saveProcessImage(processId, imageFileName, readImageResult.buffer);
 
   return new NextResponse(null, { status: 200, statusText: 'OK' });
 }
