+++ src/management-system-v2/app/api/private/[environmentId]/processes/[processId]/images/route.ts
@@ -7,9 +7,7 @@ import {
 } from '@/lib/data/legacy/_process';
 import { NextRequest, NextResponse } from 'next/server';
 import { v4 } from 'uuid';
-import stream from 'node:stream';
-import type { ReadableStream } from 'node:stream/web';
-import { fileTypeFromBuffer } from 'file-type';
+import { invalidRequest, readImage } from '../../../image-helpers';
 
 export async function GET(
   request: NextRequest,
@@ -47,23 +45,8 @@ export async function POST(
     params: { environmentId, processId },
   }: { params: { environmentId: string; processId: string } },
 ) {
-  const allowedContentTypes = ['image/jpeg', 'image/svg+xml', 'image/png'];
-
-  const contentType = request.headers.get('content-Type');
-
-  if (!contentType || !allowedContentTypes.includes(contentType)) {
-    return new NextResponse(null, {
-      status: 400,
-      statusText: 'Wrong content type. Image must be of type JPEG, PNG or SVG.',
-    });
-  }
-
-  if (!request.body) {
-    return new NextResponse(null, {
-      status: 400,
-      statusText: 'No image was given in request',
-    });
-  }
+  const isInvalidRequest = invalidRequest(request);
+  if (isInvalidRequest) return invalidRequest;
 
   const { ability } = await getCurrentEnvironment(environmentId);
 
@@ -84,41 +67,12 @@ export async function POST(
     });
   }
 
-  const reader = stream.Readable.fromWeb(request.body as ReadableStream<Uint8Array>);
-  const chunks: Uint8Array[] = [];
-  let totalLength = 0;
-  for await (const chunk of reader) {
-    if (chunk) {
-      chunks.push(chunk);
-      totalLength += chunk.length;
-      if (totalLength > 2000000) {
-        // 2MB limit
-        reader.destroy(new Error('Allowed image size of 2MB exceeded'));
-        return new NextResponse(null, {
-          status: 413,
-          statusText: 'Allowed image size of 2MB exceeded',
-        });
-      }
-    }
-  }
-  // Proceed with processing if the size limit is not exceeded
-  const imageBuffer = Buffer.concat(
-    chunks.map((chunk) => Buffer.from(chunk)),
-    totalLength,
-  );
+  const readImageResult = await readImage(request);
+  if (readImageResult.error) return readImageResult.error;
 
-  const fileType = await fileTypeFromBuffer(imageBuffer);
+  const imageFileName = `_image${v4()}.${readImageResult.fileType.ext}`;
 
-  if (!fileType) {
-    return new NextResponse(null, {
-      status: 415,
-      statusText: 'Can not store image with unknown file type',
-    });
-  }
-
-  const imageFileName = `_image${v4()}.${fileType.ext}`;
-
-  await saveProcessImage(processId, imageFileName, imageBuffer);
+  await saveProcessImage(processId, imageFileName, readImageResult.buffer);
 
   return new NextResponse(imageFileName, { status: 201, statusText: 'Created' });
 }
